+++ src/components/combo_box/combo_box.tsx
@@ -83,6 +83,10 @@ export interface _EuiComboBoxProps<T>
    */
   isClearable: boolean;
   /**
+   * When true clicking on an option will open it for editing
+   */
+  isEditable?: boolean;
+  /**
    * Disables the input
    */
   isDisabled?: boolean;
@@ -905,6 +909,7 @@ export class EuiComboBox<T> extends Component<
       id,
       inputRef,
       isClearable,
+      isEditable,
       isDisabled,
       isInvalid,
       isLoading,
@@ -998,6 +1003,21 @@ export class EuiComboBox<T> extends Component<
       );
     }
 
+    const onEdit = (option: EuiComboBoxOptionOption<T>) => {
+      if (this.state.searchValue === '') {
+        this.onRemoveOption(option);
+        this.setState({
+          searchValue: option.label,
+        });
+      }
+    };
+
+    if (this.props.isEditable) {
+      selectedOptions.forEach(
+        (option) => (option.onClick = () => onEdit(option))
+      );
+    }
+
     return (
       /**
        * Re: jsx-a11y/interactive-supports-focus
@@ -1020,6 +1040,7 @@ export class EuiComboBox<T> extends Component<
         ref={this.comboBoxRefCallback}
         role="combobox">
         <EuiComboBoxInput
+          isEditable={this.props.isEditable}
           autoSizeInputRef={this.autoSizeInputRefCallback}
           compressed={compressed}
           focusedOptionId={