+++ src/components/combo_box/combo_box_input/combo_box_input.tsx
@@ -39,6 +39,7 @@ export interface EuiComboBoxInputProps<T> extends CommonProps {
   hasSelectedOptions: boolean;
   id?: string;
   inputRef?: RefCallback<HTMLInputElement>;
+  isEditable?: boolean;
   isDisabled?: boolean;
   isListOpen: boolean;
   noIcon: boolean;
@@ -131,6 +132,7 @@ export class EuiComboBoxInput<T> extends Component<
       hasSelectedOptions,
       id,
       inputRef,
+      isEditable,
       isDisabled,
       isListOpen,
       noIcon,
@@ -163,7 +165,7 @@ export class EuiComboBoxInput<T> extends Component<
       ? selectedOptions.map((option) => {
           const { key, label, color, onClick, ...rest } = option;
           const pillOnClose =
-            isDisabled || singleSelection || onClick
+            isDisabled || singleSelection || (onClick && !isEditable)
               ? undefined
               : onRemoveOption;
           return (